backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# backend/api/dependencies.py: 6
# backend/api/v1/auth.py: 8
# backend/api/v1/games.py: 4
# backend/api/v1/matches.py: 4
# backend/api/v1/users.py: 4
# backend/crud/auth.py: 1
# backend/crud/game.py: 1,2
# backend/crud/match.py: 1
# backend/crud/movement.py: 2
# backend/crud/piece.py: 2
# backend/crud/user.py: 1
# backend/database/database.py: 1,2,3
# backend/models/game.py: 1,2,3
# backend/models/match.py: 1,2,3,4
# backend/models/movement.py: 1,2,3
# backend/models/piece.py: 1,2
# backend/models/user.py: 1,2
# backend/tests/api/v1/utils.py: 2,3
SQLAlchemy == 1.3.23

# backend/crud/auth.py: 5
# backend/crud/user.py: 5
bcrypt == 3.2.0

# backend/api/dependencies.py: 1,2,3
# backend/api/v1/auth.py: 4,5
# backend/api/v1/games.py: 1,2
# backend/api/v1/matches.py: 1,2
# backend/api/v1/users.py: 1,2
# backend/main.py: 1,2
# backend/tests/api/v1/utils.py: 1
fastapi == 0.63.0

# backend/schemas/auth.py: 1
# backend/schemas/game.py: 2
# backend/schemas/match.py: 2
# backend/schemas/movement.py: 1
# backend/schemas/piece.py: 2
# backend/schemas/user.py: 2
pydantic == 1.7.3

# backend/tests/api/v1/test_auth.py: 4
# backend/tests/api/v1/test_games.py: 4
# backend/tests/api/v1/test_users.py: 4
pytest == 6.2.2

# backend/api/dependencies.py: 4
# backend/api/v1/auth.py: 6
python_jose == 3.2.0

# Uvicorn server
uvicorn == 0.13.4
psycopg2 == 2.8.6
python-multipart == 0.0.5
